Replace Float by Double

+++ src/main/java/fr/openent/lystore/export/subventionEquipment/Market.java
@@ -87,7 +87,7 @@ public class Market extends TabHelper {
     private void setTitle() {
         for (int i = 0; i < datas.size(); i++) {
             JsonObject data = datas.getJsonObject(i);
-            totalSubv += Float.parseFloat(data.getString("totalprice"));
+            totalSubv += Double.parseDouble(data.getString("totalprice"));
         }
         excel.insertBlackTitleHeaderBorderlessCenter(0, lineNumber, ANNEXE_TEXT);
         sizeMergeRegionWithStyle(lineNumber, 0, 2, excel.blackTitleHeaderBorderlessCenteredStyle);
